+++ src/main/java/com/mca/domain/videogame/vos/VideoGameId.java
@@ -2,9 +2,9 @@ package com.mca.domain.videogame.vos;
 
 import com.mca.domain.shared.ValueObject;
 
-public class VideoGameId extends ValueObject<Integer> {
+public class VideoGameId extends ValueObject<Long> {
 
-    public VideoGameId(final int value) {
+    public VideoGameId(final Long value) {
         super(value);
     }
 }
